1-listint_len.c

+++ 0x13-more_singly_linked_lists/1-listint_len.c
@@ -1,9 +1,14 @@
 #include "lists.h"
 
+/**
+  * listint_len - gets the number of nodes in a list
+  * @h: singly linked list to print
+  * Return: number of nodes in the list
+  */
 
 size_t listint_len(const listint_t *h)
 {
-	int i;
+	size_t i;
 	
 	for (i = 0; h; i++)
 		h = h->next;
